buildflow-gcp-image-classification/requirements.txt
buildflow
cython
imageai
pandas
pillow>=7.0.0
numpy>=1.18.1
opencv-python>=4.1.2
ray
torch>=1.9.0 --extra-index-url https://download.pytorch.org/whl/cpu
torchvision>=0.10.0 --extra-index-url https://download.pytorch.org/whl/cpu
pytest==7.1.3
python-dotenv
tqdm==4.64.1
scipy>=1.7.3
matplotlib>=3.4.3
mock==4.0.3
google-cloud-bigquery
google-cloud-storage
db-dtypes
python-multipart
